+++ examples/storybook/src/resources/experimentsExample.ts
@@ -1,6 +1,8 @@
 export const EXPERIMENTS = [
   { name: 'newFormattingToolbar', scope: 'ricos' },
   { name: 'tiptapEditor', scope: 'ricos' },
+  { name: 'newVideoModal', scope: 'ricos' },
+  { name: 'newSocialAndVerticalEmbedModal', scope: 'ricos' },
   { name: 'forceIframeSandboxDomain', scope: 'ricos' },
   { name: 'useSeo', scope: 'wixRicos' },
   { name: 'tableCellsLimit', scope: 'ricos', input: [100, 200, 300, 500, 1000] },