+++ packages/ricos-editor/web/src/modals/EditorModalProvider.tsx
@@ -7,7 +7,7 @@ import React, {
   FunctionComponent,
 } from 'react';
 import mergeModalStyles from './mergeModalStyles';
-import { ModalStyles, ModalsMap } from 'wix-rich-content-common';
+import { ModalStyles, ModalsMap, AvailableExperiments } from 'wix-rich-content-common';
 import { ModalSettings } from 'ricos-common';
 import { merge } from 'lodash';
 import ReactDOM from 'react-dom';
@@ -19,6 +19,7 @@ interface Props {
   locale: string;
   ariaHiddenId?: ModalSettings['ariaHiddenId'];
   container?: HTMLElement;
+  experiments?: AvailableExperiments;
   onModalOpen: (modalProps: Record<string, unknown>) => void;
   onModalClose: () => void;
 }
@@ -91,7 +92,7 @@ export default class EditorModalProvider extends Component<Props, State> {
 
   render() {
     const { EditorModal, showModal, modalProps, modalStyles, editorModalId } = this.state;
-    const { children, ModalsMap, locale, theme, ariaHiddenId, container } = this.props;
+    const { children, ModalsMap, locale, theme, ariaHiddenId, container, experiments } = this.props;
     const childProps = merge(children.props, this.modalHandlers);
     return (
       <Fragment>
@@ -112,6 +113,7 @@ export default class EditorModalProvider extends Component<Props, State> {
                   modalsMap={ModalsMap}
                   locale={locale}
                   target={editorModalId}
+                  experiments={experiments}
                   {...modalProps}
                 />
               </Suspense>